+++ source/lib/patches.js
@@ -1,0 +1,10 @@
+
+
+if (!Object.hasOwn)
+{
+    Object.hasOwn = (object, property) =>
+    {
+        // eslint-disable-next-line no-prototype-builtins
+        return object.hasOwnProperty(property);
+    }
+}