+++ source/lib/utils/dom/index.js
@@ -1,4 +1,5 @@
 
 export { default as addDocumentListener } from './add-document-listener'
 export { default as computeStyle } from './compute-style'
+export { default as observeIntersections } from './observe-intersections'
 export { default as pageOffsets } from './page-offsets'